requirements.txt
aerich~=0.7.2
asyncpg~=0.29.0
asyncmy~=0.2.9
black~=24.2.0
cryptography~=42.0.5
fastapi~=0.110.0
jinja2~=3.1.3
pydantic~=2.6.3
pydantic-settings~=2.2.1
tortoise-orm~=0.20.0
sse-starlette~=2.0.0
uvicorn~=0.27.1